+++ CHANGELOG.md
@@ -29,6 +29,8 @@
 - Changes to SqlRS
   - Updated the integration tests to stop the Reporting Services service after
     the integration test. This is to save memory on the AppVeyor build worker.
+  - The helper function `Restart-ReportingServicesService` should no longer timeout
+    when restarting the service (issue #1114).
 
 ## 11.1.0.0